fixed firebase RTDB connection with MapBox
