+++ lib/Screen/DataPage/components/background.dart
@@ -1,10 +1,16 @@
 import 'package:flutter/material.dart';
 
 class Background extends StatelessWidget {
-  const Background({super.key});
+  final Widget child;
+  const Background({
+    super.key, 
+    required this.child
+  });
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold();
+    return SizedBox(
+      child: child
+    );
   }
 }
